Update B.4.9.0100. - Branching Logic.feature
Correcting an error in the script, previously the '##ACTION: change branching logic for all' asked the tester to change the branching logic for 'text2' and then select Yes to update branching logic for other fields. The script stated that the action should result in the updates being made to 'text2' and 'notes'. However the remaining fields which should have been edited by the change all action, were 'textbox' and 'notesbox'